Fixed assets not getting registered, renamed assets module
