+++ enginecore/enginecore/state/hardware/asset_definition.py
